changed SmallInteger to something our version of Django likes

+++ apps/mainmenu/models.py
@@ -1,6 +1,6 @@
 from __future__ import unicode_literals
 from django.db import models
-from login import Users
+from ..login.models import Users
 
 
 class CharManager(models.Manager):
@@ -14,13 +14,13 @@ class Char(models.Model):
     WIZARD = 'WI'
     name = models.CharField(max_length=60)
     cclass = models.CharField(max_length=2, choices=((WIZARD, 'Wizard'),(ROGUE, 'Rogue'),(WIZARD, 'Wizard')), default=FIGHTER)
-    strength = SmallIntegerField(default=1)
-    dexterity = SmallIntegerField(default=1)
-    intelligence = SmallIntegerField(default=1)
-    health = SmallIntegerField(default=1)
+    strength = IntegerField(default=1)
+    dexterity = IntegerField(default=1)
+    intelligence = IntegerField(default=1)
+    health = IntegerField(default=1)
     items = ManyToManyField(game.Items)
-    gold = SmallIntegerField(default=0)
-    level = SmallIntegerField(default=1)
+    gold = IntegerField(default=0)
+    level = IntegerField(default=1)
     current_room = ForeignKey(game.Rooms)
     owned_by = ForeignKey(login.users)
     slug = models.SlugField()
